+++ modules/apps/collaboration/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/upgrade/AnnouncementsWebUpgrade.java
@@ -17,9 +17,11 @@ package com.liferay.announcements.web.internal.upgrade;
 import com.liferay.announcements.web.internal.upgrade.v1_0_2.UpgradePermission;
 import com.liferay.portal.kernel.upgrade.BaseReplacePortletId;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
+import com.liferay.portal.kernel.upgrade.UpgradeException;
 import com.liferay.portal.kernel.upgrade.UpgradeStep;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
+import com.liferay.portal.upgrade.release.BaseUpgradeWebModuleRelease;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -31,8 +33,33 @@ public class AnnouncementsWebUpgrade implements UpgradeStepRegistrator {
 
 	@Override
 	public void register(Registry registry) {
+		BaseUpgradeWebModuleRelease upgradeWebModuleRelease =
+			new BaseUpgradeWebModuleRelease() {
+
+				@Override
+				protected String getBundleSymbolicName() {
+					return "com.liferay.announcements.web";
+				}
+
+				@Override
+				protected String[] getPortletIds() {
+					return new String[] {
+						"1_WAR_soannouncementsportlet", "83", "84",
+						PortletKeys.ANNOUNCEMENTS
+					};
+				}
+
+			};
+
+		try {
+			upgradeWebModuleRelease.upgrade();
+		}
+		catch (UpgradeException ue) {
+			throw new RuntimeException(ue);
+		}
+
 		registry.register(
-			"com.liferay.announcements.web", "0.0.0", "1.0.2",
+			"com.liferay.announcements.web", "0.0.0", "1.0.3",
 			new DummyUpgradeStep());
 
 		UpgradeStep upgradePortletId = new BaseReplacePortletId() {
@@ -44,6 +71,7 @@ public class AnnouncementsWebUpgrade implements UpgradeStepRegistrator {
 						"1_WAR_soannouncementsportlet",
 						PortletKeys.ANNOUNCEMENTS
 					},
+					new String[] {"83", PortletKeys.ALERTS},
 					new String[] {"84", PortletKeys.ANNOUNCEMENTS}
 				};
 			}
@@ -63,6 +91,12 @@ public class AnnouncementsWebUpgrade implements UpgradeStepRegistrator {
 		registry.register(
 			"com.liferay.announcements.web", "1.0.1", "1.0.2",
 			new UpgradePermission());
+
+		// See LPS-69656
+
+		registry.register(
+			"com.liferay.announcements.web", "1.0.2", "1.0.3", upgradePortletId,
+			new UpgradePermission(true));
 	}
 
 }