+++ modules/apps/foundation/frontend-js/frontend-js-spa-web/src/main/java/com/liferay/frontend/js/spa/web/internal/servlet/taglib/SPATopHeadJSPDynamicInclude.java
@@ -14,19 +14,27 @@
 
 package com.liferay.frontend.js.spa.web.internal.servlet.taglib;
 
-import com.liferay.frontend.js.spa.web.internal.constants.SPAWebKeys;
 import com.liferay.frontend.js.spa.web.internal.servlet.taglib.util.SPAUtil;
+import com.liferay.portal.kernel.language.Language;
 import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.servlet.taglib.BaseJSPDynamicInclude;
 import com.liferay.portal.kernel.servlet.taglib.DynamicInclude;
+import com.liferay.portal.kernel.servlet.taglib.aui.ScriptData;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.Html;
+import com.liferay.portal.kernel.util.Props;
 import com.liferay.portal.kernel.util.PropsKeys;
-import com.liferay.portal.kernel.util.PropsUtil;
+import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.WebKeys;
 
 import java.io.IOException;
+import java.io.InputStream;
 
-import javax.servlet.ServletContext;
+import java.util.HashMap;
+import java.util.Map;
+
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
@@ -48,16 +56,74 @@ public class SPATopHeadJSPDynamicInclude extends BaseJSPDynamicInclude {
 			String key)
 		throws IOException {
 
-		request.setAttribute(SPAWebKeys.SPA_UTIL, _spaUtil);
+		ScriptData scriptData = new ScriptData();
 
-		super.include(request, response, key);
+		Map<String, String> values = new HashMap<>();
+
+		ThemeDisplay themeDisplay = (ThemeDisplay)request.getAttribute(
+			WebKeys.THEME_DISPLAY);
+
+		values.put(
+			"cacheExpirationTime",
+			String.valueOf(
+				_spaUtil.getCacheExpirationTime(themeDisplay.getCompanyId())));
+
+		values.put(
+			"clearScreensCache",
+			String.valueOf(
+				_spaUtil.isClearScreensCache(request, request.getSession())));
+		values.put("excludedPaths", _spaUtil.getExcludedPaths());
+		values.put(
+			"loginRedirect",
+			_html.escapeJS(_spaUtil.getLoginRedirect(request)));
+		values.put(
+			"message",
+			_language.get(
+				_spaUtil.getLanguageResourceBundle(themeDisplay.getLocale()),
+				"it-looks-like-this-is-taking-longer-than-expected"));
+		values.put(
+			"navigationExceptionSelectors",
+			_spaUtil.getNavigationExceptionSelectors());
+		values.put(
+			"portletsBlacklist", _spaUtil.getPortletsBlacklist(themeDisplay));
+		values.put(
+			"requestTimeout", String.valueOf(_spaUtil.getRequestTimeout()));
+		values.put(
+			"timeout", String.valueOf(_spaUtil.getUserNotificationTimeout()));
+		values.put(
+			"title",
+			_language.get(
+				_spaUtil.getLanguageResourceBundle(themeDisplay.getLocale()),
+				"oops"));
+		values.put("validStatusCodes", _spaUtil.getValidStatusCodes());
+
+		scriptData.append(
+			null,
+			StringUtil.replaceToStringBundler(
+				_initTemplate, StringPool.POUND, StringPool.POUND, values),
+			"frontend-js-spa-web/liferay/init.es", ScriptData.ModulesType.ES6);
+
+		scriptData.writeTo(response.getWriter());
+	}
+
+	private static final String _initTemplate;
+
+	static {
+		try (InputStream inputStream =
+				SPATopHeadJSPDynamicInclude.class.getResourceAsStream(
+					"/META-INF/resources/init.tmpl")) {
+
+			_initTemplate = StringUtil.read(inputStream);
+		}
+		catch (IOException ioe) {
+			throw new ExceptionInInitializerError(ioe);
+		}
 	}
 
 	@Override
 	public void register(DynamicIncludeRegistry dynamicIncludeRegistry) {
 		boolean singlePageApplicationEnabled = GetterUtil.getBoolean(
-			PropsUtil.get(
-				PropsKeys.JAVASCRIPT_SINGLE_PAGE_APPLICATION_ENABLED));
+			_props.get(PropsKeys.JAVASCRIPT_SINGLE_PAGE_APPLICATION_ENABLED));
 
 		if (singlePageApplicationEnabled) {
 			dynamicIncludeRegistry.register(
@@ -67,21 +133,12 @@ public class SPATopHeadJSPDynamicInclude extends BaseJSPDynamicInclude {
 
 	@Override
 	protected String getJspPath() {
-		return "/init.jsp";
+		return null;
 	}
 
 	@Override
 	protected Log getLog() {
-		return _log;
-	}
-
-	@Override
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.frontend.js.spa.web)",
-		unbind = "-"
-	)
-	protected void setServletContext(ServletContext servletContext) {
-		super.setServletContext(servletContext);
+		return null;
 	}
 
 	@Reference(
@@ -97,8 +154,14 @@ public class SPATopHeadJSPDynamicInclude extends BaseJSPDynamicInclude {
 		_spaUtil = null;
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(
-		SPATopHeadJSPDynamicInclude.class);
+	@Reference
+	private Html _html;
+
+	@Reference
+	private Language _language;
+
+	@Reference
+	private Props _props;
 
 	private SPAUtil _spaUtil;