+++ modules/sdk/project-templates/src/main/java/com/liferay/project/templates/ProjectTemplates.java
@@ -24,10 +24,14 @@ import com.liferay.project.templates.internal.util.Validator;
 import com.liferay.project.templates.internal.util.WorkspaceUtil;
 
 import java.io.File;
+import java.io.IOException;
 
 import java.nio.file.DirectoryStream;
+import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.SimpleFileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.PosixFilePermissions;
 
 import java.util.ArrayList;
@@ -189,7 +193,21 @@ public class ProjectTemplates {
 			Files.deleteIfExists(templateDirPath.resolve("settings.gradle"));
 		}
 
-		Files.delete(templateDirPath.resolve("pom.xml"));
+		Files.walkFileTree(
+			templateDirPath,
+			new SimpleFileVisitor<Path>() {
+
+				@Override
+				public FileVisitResult preVisitDirectory(
+						Path dirPath, BasicFileAttributes basicFileAttributes)
+					throws IOException {
+
+					Files.deleteIfExists(dirPath.resolve("pom.xml"));
+
+					return FileVisitResult.CONTINUE;
+				}
+
+			});
 	}
 
 	private static void _printHelp(JCommander jCommander) throws Exception {