+++ portal-kernel/src/com/liferay/portal/kernel/upgrade/BaseReplacePortletId.java
@@ -14,14 +14,67 @@
 
 package com.liferay.portal.kernel.upgrade;
 
+import com.liferay.portal.kernel.dao.jdbc.AutoBatchPreparedStatementUtil;
+import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringUtil;
+
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author Adolfo Pérez
  */
 public abstract class BaseReplacePortletId extends BaseUpgradePortletId {
+
+	@Override
+	protected void doUpgrade() throws Exception {
+		Map<String, List<String>> renamePortletIdsMap = new HashMap<>();
+
+		for (String[] renamePortletIds : getRenamePortletIdsArray()) {
+			List<String> sourcePortletIds = renamePortletIdsMap.computeIfAbsent(
+				renamePortletIds[1], key -> new ArrayList<>());
+
+			sourcePortletIds.add(renamePortletIds[0]);
+		}
+
+		List<String> clauses = new ArrayList<>();
+
+		for (Map.Entry<String, List<String>> entry :
+				renamePortletIdsMap.entrySet()) {
+
+			List<String> sourcePortletIds = entry.getValue();
+
+			if (sourcePortletIds.size() > 1) {
+				for (int i = 0; i < (sourcePortletIds.size() - 1); i++) {
+					String portletId1 = sourcePortletIds.get(i);
+
+					for (int j = i + 1; j < sourcePortletIds.size(); j++) {
+						String portletId2 = sourcePortletIds.get(j);
+
+						clauses.add(
+							String.format(
+								"((PP1.portletId = '%s') AND (PP2.portletId " +
+									"= '%s'))",
+								portletId1, portletId2));
+					}
+				}
+			}
+		}
+
+		if (!clauses.isEmpty()) {
+			String orClauses = StringUtil.merge(clauses, " OR ");
+
+			_deleteConflictingPreferences(orClauses);
+		}
+
+		super.doUpgrade();
+	}
 
 	protected boolean hasPortlet(String portletId) throws SQLException {
 		return hasRow(
@@ -71,17 +124,42 @@ public abstract class BaseReplacePortletId extends BaseUpgradePortletId {
 		throws Exception {
 
 		if (hasResourceAction(newName)) {
-			try (PreparedStatement ps = connection.prepareStatement(
-					"delete from ResourceAction where name = ?")) {
+			StringBundler sb = new StringBundler(3);
 
-				ps.setString(1, oldName);
+			sb.append("select RA1.resourceActionId from ResourceAction RA1 ");
+			sb.append("inner join ResourceAction RA2 on RA1.actionId = ");
+			sb.append("RA2.actionId where RA1.name = ? and RA2.name = ?");
 
-				ps.execute();
+			try (PreparedStatement ps1 = connection.prepareStatement(
+					sb.toString());
+				PreparedStatement ps2 =
+					AutoBatchPreparedStatementUtil.concurrentAutoBatch(
+						connection,
+						"delete from ResourceAction where resourceActionId = " +
+							"?")) {
+
+				ps1.setString(1, oldName);
+				ps1.setString(2, newName);
+
+				ResultSet rs = ps1.executeQuery();
+
+				int deleteCount = 0;
+
+				while (rs.next()) {
+					ps2.setLong(1, rs.getLong(1));
+
+					ps2.addBatch();
+
+					deleteCount++;
+				}
+
+				if (deleteCount > 0) {
+					ps2.executeBatch();
+				}
 			}
 		}
-		else {
-			super.updateResourceAction(oldName, newName);
-		}
+
+		super.updateResourceAction(oldName, newName);
 	}
 
 	@Override
@@ -102,6 +180,41 @@ public abstract class BaseReplacePortletId extends BaseUpgradePortletId {
 		else {
 			super.updateResourcePermission(
 				oldRootPortletId, newRootPortletId, updateName);
+		}
+	}
+
+	private void _deleteConflictingPreferences(String orClauses)
+		throws SQLException {
+
+		StringBundler sb = new StringBundler(4);
+
+		sb.append("select PP1.portletPreferencesId from PortletPreferences ");
+		sb.append("PP1 inner join PortletPreferences PP2 on PP1.plid = ");
+		sb.append("PP2.plid where ");
+		sb.append(orClauses);
+
+		try (PreparedStatement ps1 = connection.prepareStatement(sb.toString());
+			PreparedStatement ps2 =
+				AutoBatchPreparedStatementUtil.concurrentAutoBatch(
+					connection,
+					"delete from PortletPreferences where " +
+						"portletPreferencesId = ?")) {
+
+			ResultSet rs = ps1.executeQuery();
+
+			int deleteCount = 0;
+
+			while (rs.next()) {
+				ps2.setLong(1, rs.getLong(1));
+
+				ps2.addBatch();
+
+				deleteCount++;
+			}
+
+			if (deleteCount > 0) {
+				ps2.executeBatch();
+			}
 		}
 	}