+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -767,8 +767,6 @@ found.
 OSGi projects, Ant plugins and themes if the `-PsnapshotIfStale` argument is
 provided.
 
-### Changed
-
 ## 2.3.0 - 2016-12-15
 
 ### Fixed
@@ -814,7 +812,17 @@ versions of Git.
 ## 2.3.8 - 2016-12-20
 
 ### Changed
-- [LPS-69802]: Update the [Liferay Gradle Plugins] dependency to version 3.0.59.
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 3.0.59.
+
+## 2.3.9 - 2016-12-21
+
+### Changed
+- [LPS-69802]: Update the [Liferay Gradle Plugins] dependency to version 3.0.60.
+
+## 2.3.10 - 2016-12-21
+
+### Changed
+- [LPS-69838]: Update the [Liferay Gradle Plugins] dependency to version 3.0.61.
 
 [Liferay CDN]: https://cdn.lfrs.sl/repository.liferay.com/nexus/content/groups/public
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
@@ -890,6 +898,7 @@ versions of Git.
 [LPS-69677]: https://issues.liferay.com/browse/LPS-69677
 [LPS-69730]: https://issues.liferay.com/browse/LPS-69730
 [LPS-69802]: https://issues.liferay.com/browse/LPS-69802
+[LPS-69838]: https://issues.liferay.com/browse/LPS-69838
 [LRDOCS-2594]: https://issues.liferay.com/browse/LRDOCS-2594
 [LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
 [LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981